CI-39: Fix problem with wrong navigation
